fmt: Move EnvFilter into its own module
